added folder and file structure for expected routes for orders and products, added commented code in server.js to use new routes, leave commented until router and model are created, moved restrictedMiddleware to globalMiddleware folder
